requirements.txt
fastapi

uvicorn

python-multipart

pymongo

passlib

bcrypt

Pyjwt

python-jose